Break up `IntermediateResult` module
