+++ src/IntermediateResult/index.ts
@@ -1,0 +1,14 @@
+export type { IntermediateResult } from "./types";
+
+export { NamePlaceholder } from "./NamePlaceholder";
+export { NativePlaceholder } from "./NativePlaceholder";
+export { NativeValue } from "./NativeValue";
+export { NodeObject } from "./NodeObject";
+export { Plural } from "./Plural";
+
+export {
+  ResultError,
+  IncompleteResultError,
+  BadNativeValueError,
+  NotANamedNodeError,
+} from "./errors";
